put skills into hook
